Introduce a wrapper for JCommander that can take json configurations as a single arg or a filename.
If the command line does NOT specify one of the new "---..." flags, then we revert to using JCommander as we always have - otherwise, we do all of the command line parsing with jackson.
I don't want to lump JCommander and Jackson into our inner-most core library's set of dependencies, but we can clean that up later.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>