Finish setting up useLocalStack, at least for Metadata Migrate.
Container templates now have support for volumes, which are done as scopes.  Volumes and volumeMounts are specified together only to reduce the number of generic parameters that needs to be dragged around.
useLocalStack has now been removed as a global and is now a parameter that's derived from the localstack: URI in the original userSchema.  That lets users mix and match, which might be more relevant if they're bringing in multiple object stores for different sources being conjoined into one.  Now we have the pattern there.
The MigrationConfigTransformer now replaces the localstack hostname with the IP address to placate the S3 client in RFS & MetadataMigrate.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>