Handle passing args to RFS.
I had to update the JsonCommandLineParser to make sure that it calls all of the designated converters before filling in the values in the arguments class.  There may be some interop issues between argo and the command line args, especially for things like arrays - but we can address those as we find them.  I deleted some dead converters too.

The command line options themselves are now being constructed in one step that invokes another that is responsible for setting up the replicaset.  When I tried to flatten that, the dictionary and its construction within the manifest was throwing 3 different errors depending upon which strategy that I was using.  Keeping all argo expressions outside of manifest construction seems wise.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>