Package the transform and initialization steps into a single script that is deployed on the migration console.
Add a makeBundle.js script to make a bundle.  Since some of the packages can't be bundled with esbuild, I'm packaging those into their own directory alongside the bundled index.js file.
On the docker front, I've had to remove the rule to ignore build directories because the node_modules DO have some build occurrences that are critical for the code to run.
The dockerfile itself now installs node 22 onto the console image (NOT a builder image) since we need that to run the script.
I've changed the way that gradle stages the config-processor since it's significantly simpler now with the bundled release.
Lastly, the make-templates npm script is run from gradle so that the output goes directly into the nodeStaging area, where it is picked up by the image build alongside the configProcessor.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>