WIP - Snapshot changes & improvements + DSL tweak.
Create a new snapshot every run, deleting none. The idea is that when doing a subsequent snapshot, the files already present from prior ones that are still present will reduce the time to take the new snapshot.
I've also stopped invoking create snapshot from the python CLI & instead run it directly on the migration console, just as that happens for metadata migrate. That's what made running a deletion beforehand more taxing (it would have been two sequential container invocations).
Lastly, doing that lets us easily map create snapshot arguments into command line parameters.
There's still a change in-progress. `"when": "'{{inputs.parameters.alreadyDefinedName}}' == ''"` has been replace with `when: true` because I've refactored how names are created. At that point of the workflow, the already defined (BYO) snapshot name isn't an input, but rather an output of a previous step. I've started laying the groundwork for when and loop parameters to be set via a callback using context that's similar to the context passed for register.
Changes also had to be made to the expressions file to handle a few more union types.
I've also further refactored some cluster-parameters handling code to generalize it across sources and targets.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>