Do the last bits of refactoring so that when and loop values can be specified with a callback.
I need to make a sweep over this codebase to come up with some more consistent generic parameter names.  Right now, it's confusing to know which ones are at which levels of specialization.
createOrGetSnapshot has been updated so that createSnapshot only runs when the prior step indicated that it should.
I haven't tested the live workflows yet to confirm that they're behaving as I want them to.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>