Get create & check snapshot working partially with http creds from a secret.
Creating the snapshot can successfully rig the environment variable creds now.  The username/password are still persisted within the user config and passed through to the services.yaml so that the migration console that gets loaded to check for completion of a snapshot can succeed.
Snapshot creation now has a repoName that's set by default to the same value that's used by the migration console.  Eventually, we'll want to map that to the right spot within the console - but for now, when the user doesn't specify it, we'll use the same value and everything works.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>