Restructure the main method of RFS so that a 'continuous-mode' can be supported.
In continuous mode, if runner completed successfully, then it loops around, grabs a new work item, and works on it just as it did for the last work item.
Successfully means no exceptions and that the entire work was finished.  Lease expiration is still intact and reset at the onset of each loop.
lucene and s3 downloaded files are cleared at the end of each loop (continuous mode or not).  That file removal is done via a utility that was moved out of test code an into its own utility class, hence the reason for the test updates.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>