Assorted ugfixes to get a migration working again and removed "preApproval" step.
cliRouter now awaits for main() calls to return and each of the files with main functions now have an extra guard that they only run automagically when NOT being run from the cliRouter bundle.  As the code was, the main functions were being called twice.
Omit the approval flags, in addition to the preexisting ones, when consing the json to pass to metadata and RFS.
Add a utility method to userSchemas to confirm that every scalar value with a default also has an optional & in that order.  Some helper code in the migrations workflow templates package assumes that if a value has a default, it is an optional and without a default, it's required.  That wasn't the case for one definition, so now I'm checking for that with validateOptionalDefaultConsistency.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>