Bugfixes.
Source endpoint (and all endpoints) now are only added to the cluster settings map if it was present.
There's a NODEJS env var set in the container to use node-22 so that the TS programs can work.  The helper shell script to convert/initialize the workflow has been updated.
The npm script for check-and-make-templates now directly calls npx rather than deferring to npm again so that `-- --outputDirectory` will work correctly.  Without that, the templates weren't actually landing on the image.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>