Relax the regex for the source clusters' endpoints to include the empty string.
That caused me to need to split the regex for the target and cluster since the target requires a well formed http url.  The source has an optional, which means for argo, it needs a default value which also needs to be a string.  That's how I arrived at this regex (I'd rather have a better way to deal with optionals, but we don't for now)

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>