Wire up the CONFIG_PROCESSOR_DIR as the gradle output.
The pytest will automatically run gradle to build/stage the nodejs configprocessor scripts, which the workflow integ tests can use.
Most of the contents of the resources/scripts directory were already not being used.  sample.yaml still was, and that's staying since it's dependent upon the docker build to pull in the schema (which is dynamically set).

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>