Update the AWS README to use a new minified Cfn template...
Elsewhere, I had made minor changes to the Cfn template that made it slightly bigger, which caused the `aws cloudformation deploy...` command to error out because the template was too big.  Cfn required that any templates over ~51K bytes be pulled from S3.
Rather than creating an extra step and extra infrastructure, I decided to minify the template.  I added a minify task to gradle so that it can be done pretty smoothly by a dev but not affect anything else in our build/distributions.  It might be nice to consider minifying these for users so that if they download the template, they'd be able to apply the file, but we can take that up as a separate task.  Right now, I'm focused on improving the developer documentation story.
As I added the minify task, I was getting cache-config issues, so I reworked it with the help of multiple chatbots.  This is the way forward for gradle tasks - the injection is to let it cache more configs.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>