Changes to GitHub Action release drafter to build images for both arm64 AND amd64
release-drafter has changed considerably (in addition to act guards - see below).  We used to have 3 steps to publish images - build images; tag them; push them.  We do those now in one step ("Build and Push to DockerHub").  However, so that we can do cross-platform builds, we now run the :buildImages:buildImagesToRegistry target.  That requires us to get buildkit configured.  GitHub actions supports this, but does require a couple new steps to do that.  Since all of the docker image steps are together, some of the credentials work has also been shifted around.
Lastly, it's worth noting that we build the two different sets of images from one GHA step but with two distinct gradle commands.  That made it a much simpler gradle change - though it would be nice if gradle could easily build them all in one invocation.

The buildImages/build.gradle file and supporting files from buildSrc have been updated to take in a repoName since dockerhub has a separate repo for each image.  Flags were also added as inputs to try to get better dependency management to determine when rebuilds are required.

I removed the gradle setup from the CI environment and have replaced all the other occurrences of 'gradle' with './gradlew'.  The setup-gradle action isn't supported with act (because it's an unsupported compound object - something like that).  We shouldn't need gradle anyway as we've got it packaged in our repo anyway.
I've also tried adding the 4 orchestrationSpecs npm packages to run node tests on.  We'll run some of them as part of migration-console builds but it would be nice to run them here for visibility and completeness.

I tested some of the release-drafter file/workflow with act (https://nektosact.com/).  In order to get that to work, I had to change some if statements through the workflow definition.  They seemed sensible to keep in place to support more local development.  I tested with...
```
act workflow_dispatch \
  -W .github/workflows/release-drafter.yml \
  -j draft-a-release \
  -P ubuntu-22.04=ghcr.io/catthehacker/ubuntu:act-22.04 \
  --container-architecture linux/arm64 -e act.yml
```
That tried to push to the dockerhub opensearchstaging, which I don't have permissions for, but I could see that we made it up to that point.  I also tested the gradle script against my own registry, where I DO have access to push images and tags.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>