add a backup node pool for the migration console...
in case t2/t3 instances aren't available.  I want to be careful on disruption budgets - if we're in the general node pool, being bumped to go to a the dedicated one wouldn't be so bad.
Also note that I had to put some slack space in the pool - more than just what the console requires since there are daemonset nodes on the host too.  That should be fixed.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>