Set up multi-platform support for builImages so that we can properly build AND PUSH tags for both images.
The old code was building for each separately and pushing tags separately.  Each push was ONLY for one architecture, so it knocked out the previous labels that were present from the old push.
No, we only need one gradle command instead of two to do builds an pushes...
BUT - we support basically one of three modes within a single gradle execution.  Either multi-platform, arm64 or amd64.  Each of those is mutually exclusive for targets.
We have aggregate targets for jibAll, buildKitAll, and buildImagesToRegistry - each of those will handle BOTH amd and arm.  To do ONLY one, use the _amd64 or _arm64 suffix (e.g. jibAll_amd64) - warning that pushing just one image will leave you exposed if a different platform tries to pull it.  There is a check in place to throw when multiple conflicting targets are specified.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>