Hardcode the resources values for the migration console in the helm chart.
There are too many other values that would need to shift too - it isn't worth giving the user that amount of rope to trip on.
I've also changed the list of t3 instances to just be a class of 't' instances.  I have the resource limits there for the pod, so we should settle in on a medium or greater node - shouldn't matter if it's amd or arm or what generation.
Notice that now the dedicated node pool is always on for EKS clusters - though we still fall back to the general one if the dedicated pool doesn't have anything that works.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>