WIP - adding support across the board so that parallel snapshots 'just work' for a user.
Argo semaphores and concurrency primitives have been added to the workflow build model and to dag, task, container, and resource builders.
Sprig regex operations functions are now supported by the expressions module.
Snapshot repos are no longer specified as a single repo on a source.  Now they're specified as a record for each source (optionally, when snapshotting is required).
The overall migration object schema now has a refinement on it to verify that referenced repo names are actual keys for the source's repo map.
To enable the refinement, I had to remove the deepStrict helper method, which wasn't reliable and wasn't preserving those refinements.  Now I do strict checks in a separate second pass after running the safeParse.
init.sh and traces of it have been deleted - all initialization is done outside of the workflow now.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>