Move the responsibility of configuring the snapshot semaphores to outside the argo workflow.
Approval and snapshot semaphore configs are now both initialized from the createMigrationWorkflowFromUserConfiguration.sh script - but from files that were emitted by the migration initializer TS script.  The initializer script now outputs several files into a temporary directory (passed as an argument) and the caller picks those up and applies them with kubectl.  The migration workflow configuration is also output as a file and the script splices that file into the workflow kubectl command rather than pulling it from stdout.
I've adjusted the argo schema to include semaphore names/keys so that concurrency decisions can be made upfront.  That reduces the complexity of the argo workflow and allows the process kicking that off to have more control over the concurrency of snapshots being taken.

Signed-off-by: Greg Schohn <greg.schohn@gmail.com>